fix using const val

+++ migrations/2023-05-05-005306_items_table/up.sql
@@ -6,9 +6,9 @@ CREATE TABLE item (
   price bigint NOT NULL,
   negotiable boolean NOT NULL DEFAULT false,  
   owner_id bigint NOT NULL  REFERENCES users(id),
-  favorite_count bigint NOT NULL DEFAULT 0,
-  message_count bigint NOT NULL DEFAULT 0,
-  seen_count bigint NOT NULL DEFAULT 0,
+  favorite_count int NOT NULL DEFAULT 0,
+  message_count int NOT NULL DEFAULT 0,
+  seen_count int NOT NULL DEFAULT 0,
   created_at timestamp with time zone DEFAULT now() NOT NULL,
   updated_at timestamp with time zone DEFAULT now() NOT NULL,
   deleted_at timestamp with time zone DEFAULT NULL
